Fixed some problems with the tests caused by the int to size_t conversion

+++ cantera18/branches/scons/Cantera/src/equil/vcs_solve_TP.cpp
@@ -3623,7 +3623,7 @@ namespace VCSnonideal {
     }
  
     size_t iph = m_phaseID[kspec];
-    size_t irxn = kspec - m_numComponents;
+    int irxn = int(kspec) - int(m_numComponents);
     vcs_VolPhase *VPhase = m_VolPhaseList[iph];
     int phaseExist = VPhase->exists();
